refactor: consolidate MultiTexture types into multiTexture.pure.ts and delete types file
