+++ packages/public/@babylonjs/core/package.json
@@ -77,6 +77,7 @@
         "Engines/WebGPU/Extensions/*types.js",
         "Layers/*types.js",
         "Materials/*types.js",
+        "Materials/Textures/*types.js",
         "Meshes/*types.js",
         "Misc/*types.js",
         "PostProcesses/RenderPipeline/Pipelines/*types.js",
@@ -420,17 +421,13 @@
         "Materials/Textures/Procedurals/proceduralTexture.js",
         "Materials/Textures/baseTexture.js",
         "Materials/Textures/baseTexture.polynomial.js",
-        "Materials/Textures/baseTexture.polynomial.types.js",
         "Materials/Textures/colorGradingTexture.js",
-        "Materials/Textures/colorGradingTexture.types.js",
-        "Materials/Textures/cubeTexture.types.js",
         "Materials/Textures/envCubeTexture.js",
         "Materials/Textures/equiRectangularCubeTexture.js",
         "Materials/Textures/exrCubeTexture.js",
         "Materials/Textures/hdrCubeTexture.js",
         "Materials/Textures/mirrorTexture.js",
         "Materials/Textures/multiTexture.js",
-        "Materials/Textures/renderTargetTexture.types.js",
         "Materials/Textures/texture.js",
         "Materials/colorCurves.js",
         "Materials/effectRenderer.js",
